댓글 추가

+++ src/main/java/com/example/nightclub/DTO/ContentRequestDto.java
@@ -11,7 +11,7 @@ public class ContentRequestDto {
     private Long id;
     private String content;
     private String userid;
-    private Long post_id;
+    private int post_id;
 
     public Content toEntity() {
         return Content.builder()  // 빌더 패턴을 사용하여 엔티티 생성
